requirements.txt
streamlit
pandas
matplotlib
seaborn
urlextract
emoji
numpy
wordcloud
plotly
fpdf2
textblob